Created the profile screen and the profile button by importing react icons
The set up of the profile page is complete, however, I will figure out how to make the profile design appear differently when a manger user logs in compared to when an employee user logs in. Additionally, I will have to understand how the information will be populate from the database.